Filter anime according to preferences

Filter the anime list depending on different variables. These variables must be included in the data set.

newcommer_recom(data, age, gender, freetime)

Arguments

data
data containing the list of anime with the variables being description for each anime

age
age given by the person

gender
theme of the anime. For instance if you like action anime, put "action"

freetime
free time that the person has to use in order to watch the anime

Value

Return a filtered table with the anime depending on the inputs

Author

Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger

Examples

#loading the data from the package
anime <- tibble::tibble(ProjectG5::anime)

#filtering the table to make propositions to the user through the function
newcommer_recom(anime, 15, "Sport", 30)
#> # A tibble: 315 x 7
#>    Name                      Genders   Type  Episodes Duration Rating Popularity
#>    <chr>                     <chr>     <chr> <chr>       <int> <chr>       <dbl>
#>  1 Haikyu!!                  Comedy, ~ TV    25             24 PG-13~         50
#>  2 Kuroko's Basketball       Comedy, ~ TV    25             24 PG-13~        104
#>  3 Kuroko's Basketball 2     Comedy, ~ TV    25             24 PG-13~        185
#>  4 Free! - Iwatobi Swim Club Slice of~ TV    12             24 PG-13~        189
#>  5 Yuri!!! On ICE            Comedy, ~ TV    12             23 PG-13~        191
#>  6 Kuroko's Basketball 3     Comedy, ~ TV    25             24 PG-13~        232
#>  7 Haikyuu!!: To the Top     Comedy, ~ TV    13             24 PG-13~        276
#>  8 Fighting Spirit           Comedy, ~ TV    75             23 PG-13~        345
#>  9 Chihayafuru               Drama, G~ TV    25             22 PG-13~        389
#> 10 Free! - Eternal Summer    Slice of~ TV    13             24 PG-13~        450
#> # ... with 305 more rows